requirements.txt
keboola.component
snowflake-connector-python==2.3.10